feat(): routing platform independence (polishing)
